+++ .gitignore
@@ -3,3 +3,5 @@ obj
 published
 release
 src/service/state.json
+/.vs
+/ChessBuddies.sln
